medical-text-web-model/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
joblib==1.3.2
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
imbalanced-learn==0.11.0 
python-dotenv==1.0.1
Flask-Login==0.6.3
Flask-PyMongo==2.3.0 